backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
supabase==2.3.4 